display_update updates X11 authentication
Solves issue with same display 'xauth list' entry but actual
different displays and authentication tokens.

Issue example:
In host machine A
1. Enter tmux session
2. Enter distrobox container
    * X11 authentication works good
3. Detach tmux session

In host machine B
1. SSH to host machine A
2. Enter distrobox container
    * X11 authentication does NOT work anymore

This is because in both examples, the same display e.g.
    DISPLAY=localhost:10.0
is used. Whereas the 2 different machines use different tokens,
but as the :10 display already exists in 'xauth list' it doesn't
update.